test: moved excess_length_strat test to test_data.py
